requirements.txt
numpy==1.22.4
Pillow==8.1.0
Pillow==9.2.0
Pillow==11.0.0
scikit_image==0.18.2
scipy==1.7.3
torch==1.10.1+cu113
torchvision==0.11.1+cu113
